+++ web/documentserver-example/java-spring/src/main/java/com/onlyoffice/integration/sdk/service/ConfigServiceImpl.java
@@ -215,6 +215,7 @@ public class ConfigServiceImpl extends DefaultConfigService implements ConfigSer
                 .id(String.valueOf(appUser.getId()))
                 .name(appUser.getName())
                 .group(appUser.getGroup().getName())
+                .roles(appUser.getRoles())
                 .build();
 
         if (appUser.getAvatar()) {